+++ src/basic/components/AddToCartButton.js
@@ -1,0 +1,7 @@
+export default function AddToCartButton({ text }) {
+  return /* HTML */ `
+    <button id="add-to-Cart" class="bg-blue-500 text-white px-4 py-2 rounded">
+      ${text}
+    </button>
+  `;
+}
