+++ src/basic/components/CartContainer.js
@@ -1,0 +1,3 @@
+export default function CartContainer({ children }) {
+  return /* HTML */ ` <div class="bg-gray-100 p-8">${children}</div> `;
+}
