+++ src/basic/components/CartHeader.js
@@ -1,0 +1,3 @@
+export default function CartHeader({ text }) {
+  return /* HTML */ `<h1 class="text-2xl font-bold mb-4">${text}</h1>`;
+}
