+++ src/basic/components/CartProductList.js
@@ -1,0 +1,3 @@
+export default function CartProductList() {
+  return /* HTML */ `<div id="Cart-items"></div>`;
+}
