+++ src/basic/components/CartTotal.js
@@ -1,0 +1,5 @@
+export default function CartTotal() {
+  return /* HTML */ `
+    <div id="Cart-total" class="text-xl font-bold my-4"></div>
+  `;
+}
