+++ src/basic/components/ProductSelector.js
@@ -1,0 +1,5 @@
+export default function ProductSelector() {
+  return /* HTML */ `
+    <select id="product-select" class="border rounded p-2 mr-2"></select>
+  `;
+}
