+++ src/basic/components/StockStatus.js
@@ -1,0 +1,5 @@
+export default function StockStatus({ text }) {
+  return /* HTML */ `
+    <div id="stock-status" class="text-sm text-gray-500 mt-2">${text}</div>
+  `;
+}
