+++ src/basic/components/index.js
@@ -1,0 +1,7 @@
+export { default as AddToCartButton } from './AddToCartButton';
+export { default as CartContainer } from './CartContainer';
+export { default as CartHeader } from './CartHeader';
+export { default as CartProductList } from './CartProductList';
+export { default as CartTotal } from './CartTotal';
+export { default as ProductSelector } from './ProductSelector';
+export { default as StockStatus } from './StockStatus';
