+++ src/basic/lib/createObserver.js
@@ -1,0 +1,7 @@
+export function createObserver() {
+  const listeners = new Set();
+  const subscribe = (fn) => listeners.add(fn);
+  const notify = () => listeners.forEach((listener) => listener());
+
+  return { subscribe, notify };
+}
